Move some classes into other packages.
