+++ bundles/wps/wps-service/src/main/java/org/orbisgis/wpsservice/WpsServerImpl.java
@@ -6,8 +6,11 @@ import net.sourceforge.cobertura.CoverageIgnore;
 import org.orbisgis.corejdbc.DataSourceService;
 import org.orbisgis.wpsservice.controller.execution.DataProcessingManager;
 import org.orbisgis.wpsservice.controller.execution.ProcessExecutionListener;
+import org.orbisgis.wpsservice.controller.execution.ProcessWorker;
 import org.orbisgis.wpsservice.controller.process.ProcessIdentifier;
 import org.orbisgis.wpsservice.controller.process.ProcessManager;
+import org.orbisgis.wpsservice.controller.utils.Job;
+import org.orbisgis.wpsservice.model.JaxbContainer;
 import org.orbisgis.wpsservice.utils.ProcessTranslator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
