+++ bundles/wps/wps-service/src/main/java/org/orbisgis/wpsservice/controller/execution/ProcessWorker.java
@@ -1,9 +1,8 @@
-package org.orbisgis.wpsservice;
+package org.orbisgis.wpsservice.controller.execution;
 
 import net.opengis.wps._2_0.DescriptionType;
 import net.opengis.wps._2_0.ProcessDescriptionType;
-import org.orbisgis.wpsservice.controller.execution.DataProcessingManager;
-import org.orbisgis.wpsservice.controller.execution.ProcessExecutionListener;
+import org.orbisgis.wpsservice.controller.utils.Job;
 import org.orbisgis.wpsservice.controller.process.ProcessIdentifier;
 import org.orbisgis.wpsservice.controller.process.ProcessManager;
 import org.slf4j.LoggerFactory;
