+++ bundles/wps/wps-service/src/main/java/org/orbisgis/wpsservice/controller/utils/Job.java
@@ -1,4 +1,4 @@
-package org.orbisgis.wpsservice;
+package org.orbisgis.wpsservice.controller.utils;
 
 import net.opengis.wps._2_0.ProcessDescriptionType;
 import org.orbisgis.wpsservice.controller.execution.ProcessExecutionListener;
